requirements.txt
astroid==2.4.2
certifi==2020.6.20
chardet==3.0.4
dnspython==2.0.0
filelock==3.0.12
idna==2.10
isort==5.6.4
lazy-object-proxy==1.4.3
mccabe==0.6.1
py3-validate-email==0.2.10
pylint==2.6.0
requests==2.24.0
six==1.15.0
toml==0.10.1
urllib3==1.25.11
wrapt==1.12.1
